requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchvision==0.15.2+cu118
numpy
tqdm
PyYAML
Pillow